requirements.txt
aws-cdk-lib>=2.100.0
constructs>=10.0.0
boto3>=1.34.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
flask>=2.3.0
flask-cors>=4.0.0
